+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [0.26.3] - 2025-04-11
+
 ### Fixed
 
 - Path to the `nrfutil` wasn't correctly set when using `EAST_DONT_USE_TOOLCHAIN_MANAGER=1`.
@@ -756,7 +758,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Docker scripts for building and running docker containers, for development
   purposes.
 
-[unreleased]: https://github.com/IRNAS/irnas-east-software/compare/v0.26.2...HEAD
+[unreleased]: https://github.com/IRNAS/irnas-east-software/compare/v0.26.3...HEAD
+[0.26.3]: https://github.com/IRNAS/irnas-east-software/compare/v0.26.2...v0.26.3
 [0.26.2]: https://github.com/IRNAS/irnas-east-software/compare/v0.26.1...v0.26.2
 [0.26.1]: https://github.com/IRNAS/irnas-east-software/compare/v0.26.0...v0.26.1
 [0.26.0]: https://github.com/IRNAS/irnas-east-software/compare/v0.25.3...v0.26.0